requirements.txt
# Job Application Agent Dependencies
anthropic>=0.40.0
playwright>=1.45.0
aiosqlite>=0.20.0
pydantic>=2.0.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
jinja2>=3.1.0
requests>=2.32.0
aiohttp>=3.9.0

# Kilonova Detection Notebooks
numpy
pandas
matplotlib
scipy
jupyter
